Use correct return types and values.

+++ src/rtk/io.h
@@ -21,11 +21,11 @@ int rtk_io_fwrite(void * fp, void * buffer, int n);
 
 /* special I/O functionality for easier loading of MIDI file data */
 int rtk_io_fgetc(void * fp);
-int rtk_io_mgetw(void * fp);
+short rtk_io_mgetw(void * fp);
 long rtk_io_mgetl(void * fp);
 int rtk_io_fputc(int c, void * fp);
 int rtk_io_mputw(int w, void * fp);
-long rtk_io_mputl(long l, void * fp);
+int rtk_io_mputl(long l, void * fp);
 
 /* driver functions */
 int rtk_set_io_driver(RTK_IO_DRIVER * dp);
